[ keymap ] Map n/N to mnnzz/mnNzz
mnnzz/mnNzz means:

- Set mark to n

- Repeat the latest "/" or "?"

- Center the cursor

And then we can go back to previous cursor position by type 'n.

This map is useful when in the middle of viewing a long code, and
accidentally type n/N and throwed away to another part of code, we can
easily go back to the last cursor position by type 'n.