[ plugin/jdtls ] Use the `config` spec to create `AttachJDTLS` event
This is a follow up for 7af60b8, the actual JDTLS configuration is
called inside `AttachJDTLS` event.

This commit also use `ft` spec for lazy load instead of `module` spec,
the plugin will be loaded on `.java` filetype, so we can be sure that
all the code under the `config` spec can access the plugin.