[sugar.highlight] Add `get` method to returns the attributes as table
This method returns the highlight attributes as table where each keys is
a function, with this approach, we can lazy load attribute values.

For example, this is without this method

```lua
local c = require("colors")
local hi = require("sugar.highlight")

hi.colorscheme({
  { "StatusLine", fg = c.text, bg = c.bg_alt },
  { "VertSplit", fg = hi.bg("StatusLine") },
})
```

We overriding `StatusLine` and then used `StatusLine`'s `bg` for
`VertSplit`'s `fg`. we expected that `hi.bg` return the same
value as we have overridden, but it doesn't, what we get is
the old `StatusLine` as it hasn't been overridden.

With this method, we can do it like this

```lua
{ "VertSplit", fg = hi.get("StatusLine").bg },
```

Now, the `VertSplit`'s `fg` is a function, so we can call it when
the `StatusLine` has been overridden and since `colorscheme`
method applies the highlights sequentially, we can ensure
that the next group gets the overridden value.