config.event: Check for normal buffer
The implementation of the autocmds intended for normal buffers,
like `cursorline_on_current_window`, I don't want to apply that autocmd
in a window designed to have cursors, e.g. LSP symbol window
that shows us the current symbols based on the current cursor position
in the editing buffer window.