config.plugin.colorscheme.gruvbox: Fix light mode
Color with a `bright_` prefix for dark mode and its light mode
variant is the one with `faded_` prefix, avoid using them directly
unless the colors look contrast enough in both modes.

The `p` table acts as an intermediate variable that conditionally
swicthes the palette from `bright_` to `faded_`.

The `dark*` and `light*` colors are better accessed by extracting
the `Gruvbox*` group.