plugin.lualine: Use `StatusLineM%Mode%` highlight for section `a`
This change fixes an issue where changing the bg color of
`StatusLineM%Mode%` in the colorscheme setup file doesn't affect
the `a` section, this is due to the `theme` module harcoding the section
highlight by extracting the fg color from `StatusLineM%Mode%` and
the bg color from `StatusLine`.

With this change, the colorscheme setup file responsible for the
appearance of the `a` section highlight.